Added Swig into the unity project. Completed the server files from the cpp tutorial for raknet into c#. The unity editor crashes when the project is run.

+++ DLL/RakNetC#/RaknetWithSwig/NathanNetServer/NathanNetServer/NathanNetServer/GameMessages.cs
@@ -36,7 +36,7 @@ namespace NathanNetServer
                 {
                     switch (packet.data[0])
                     {
-                        case RakNet.DefaultMessageIDTypes.ID_NEW_INCOMING_CONNECTION:
+                        case (int)RakNet.DefaultMessageIDTypes.ID_NEW_INCOMING_CONNECTION:
                             
                             break;
                     }
